# target/generated-sources/swagger/apifirst/com/apifirst/controllor/GetApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.apifirst.controllor;

import com.apifirst.model.Carts;
import com.apifirst.model.ErrorModel;

import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import java.util.List;
import javax.validation.constraints.*;
import javax.validation.Valid;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-11-25T00:07:42.828-06:00")

@Api(value = "get", description = "the get API")
public interface GetApi {

    @ApiOperation(value = "Cart Service - Get the existing cart", notes = "Get cart service - Get the existing carts for  user", response = Carts.class, tags={ "CartService", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Get cart Responses", response = Carts.class),
        @ApiResponse(code = 400, message = "", response = Void.class),
        @ApiResponse(code = 500, message = "General server error", response = ErrorModel.class),
        @ApiResponse(code = 200, message = "Unexpected error", response = ErrorModel.class) })
    
    @RequestMapping(value = "/get/{userId}",
        produces = { "application/json" }, 
        consumes = { "application/x-www-form-urlencoded" },
        method = RequestMethod.GET)
    default ResponseEntity<Carts> getCartService(@ApiParam(value = "User Id",required=true ) @PathVariable("userId") String userId) {
        // do some magic!
        return new ResponseEntity<Carts>(HttpStatus.OK);
    }

}